+++ RetailCoder.VBE/App.cs
@@ -3,7 +3,7 @@ using System.Drawing;
 using System.Globalization;
 using System.Windows.Forms;
 using Microsoft.Vbe.Interop;
-using Rubberduck.FindAllImplementations;
+using Rubberduck.Navigations;
 using Rubberduck.Inspections;
 using Rubberduck.Parsing;
 using Rubberduck.Parsing.VBA;
