+++ RetailCoder.VBE/UI/RefactorMenu.cs
@@ -5,7 +5,7 @@ using System.Runtime.InteropServices;
 using System.Windows.Forms;
 using Microsoft.Office.Core;
 using Microsoft.Vbe.Interop;
-using Rubberduck.FindAllImplementations;
+using Rubberduck.Navigations;
 using Rubberduck.Parsing;
 using Rubberduck.Parsing.Symbols;
 using Rubberduck.Properties;
